+++ test/plugin/remote_url_test.rb
@@ -30,7 +30,7 @@ describe Shrine::Plugins::RemoteUrl do
 
       it "is not defined for entity attachments" do
         @shrine.plugin :model
-        @model_class.include @shrine::Attachment.new(:file, type: :entity)
+        @model_class.include @shrine::Attachment.new(:file, model: false)
 
         refute @model_class.method_defined?(:file_remote_url=)
       end
@@ -48,7 +48,7 @@ describe Shrine::Plugins::RemoteUrl do
 
       it "is not defined for entity attachments" do
         @shrine.plugin :model
-        @model_class.include @shrine::Attachment.new(:file, type: :entity)
+        @model_class.include @shrine::Attachment.new(:file, model: false)
 
         refute @model_class.method_defined?(:file_remote_url)
       end